requirements.txt
wget~=3.2
cffi~=1.15.0
pillow~=5.3.0
zlib~=1.2.12
numpy~=1.19.5
setuptools~=58.0.4
matplotlib~=3.3.4
python-dateutil~=2.8.2
cycler~=0.11.0
pyparsing~=3.0.9
kiwisolver~=1.3.1
opencv-python~=4.5.5.64
moviepy~=1.0.3
nls~=0.0.1